Backend compiler improvements
Initialization writes out the tmp/elm.json file from a hard-coded JSON value
which saves having to find one in the filesystem. Running 'elm make' using the
elm.json from the examples directory complains because there is no 'src'
directory so this also gets round that. The two could theoretically get out of
sync but this isn't a big problem for now.

Fixed incorrect path in pathToArtifact function - the output path seems to have
changed in 0.19, which caused errors deleting the .elmi and .elmo files.